fix phpstan issues

+++ app/Http/Controllers/ProfileController.php
@@ -26,13 +26,13 @@ class ProfileController extends Controller
      */
     public function update(ProfileUpdateRequest $request): RedirectResponse
     {
-        $request->user()->fill($request->validated());
+        $request->user()?->fill($request->validated());
 
-        if ($request->user()->isDirty('email')) {
+        if ($request->user()?->isDirty('email')) {
             $request->user()->email_verified_at = null;
         }
 
-        $request->user()->save();
+        $request->user()?->save();
 
         return Redirect::route('profile.edit')->with('status', 'profile-updated');
     }
@@ -50,7 +50,9 @@ class ProfileController extends Controller
 
         Auth::logout();
 
-        $user->delete();
+        if ($user) {
+            $user->delete();
+        }
 
         $request->session()->invalidate();
         $request->session()->regenerateToken();
